add tankData file seperate and imported
