+++ index.html
@@ -54,7 +54,7 @@
       </div>
     </div>
   </div>
-  <script src="./main.js"></script>
+  <script src="./main.js" type="module"></script>
 </body>
 </html>
