+++ main.js
@@ -1,38 +1,16 @@
 
+import tankData  from './tankData.js'
+
 function toggleMenu() {
     const menu = document.getElementById("menu");
     const hamburger = document.querySelector(".hamburger");
   
     menu.classList.toggle("active");
-    hamburger.innerHTML = menu.classList.contains("active") ? "&#x2715;" : "☰"; // Toggle icon
+    hamburger.innerHTML = menu.classList.contains("active") ? "&#x2715;" : "☰";
   } 
     
     function calculateFoam() {
-      const tankData = {
-        11: { diameter: 95, capacity: 50000, material: "Crude Oil" },
-        12: { diameter: 219, capacity: 300000, material: "Jet A" },
-        13: { diameter: 219, capacity: 300000, material: "Crude Oil" },
-        14: { diameter: 219, capacity: 300000, material: "Crude Oil" },
-        20: { diameter: 200, capacity: 223000, material: "Low Sulfur VTB/LCO" },
-        22: { diameter: 134, capacity: 100000, material: "Diesel/Jet" },
-        23: { diameter: 219, capacity: 300000, material: "HVGO" },
-        25: { diameter: 252, capacity: 452000, material: "Hih Sulfur Vac Bottoms" },
-        33: { diameter: 52, capacity: 15000, material: "Bunker/VTB" },
-        35: { diameter: 134, capacity: 100000, material: "Jet" },
-        36: { diameter: 173, capacity: 200000, material: "Diesel Fuel #2" },
-        37: { diameter: 134, capacity: 100000, material: "Light Cycle Oil/Jet A" },
-        42: { diameter: 85, capacity: 40000, material: "Asphalt" },
-        45: { diameter: 134, capacity: 100000, material: "Napatha"},
-        51: { diameter: 60, capacity: 20000, material: "Asphalt" },
-        64: { diameter: 134, capacity: 100000, material: "UnLeaded Gasoline" },
-        65: { diameter: 134, capacity: 100000, material: "UnLeaded Gasoline" },
-        66: { diameter: 134, capacity: 100000, material: "Sour Naphtha and Kerosene" },
-        67: { diameter: 160, capacity: 175000, material: "Napatha"},
-        70: { diameter: 85, capacity: 35000, material: "Butane" },
-        80: { diameter: 85, capacity: 40000, material: "Propane" },
-      };
-
-      const tankNumber = parseInt(document.getElementById("tankNumber").value);
+      const tankNumber = parseInt(document.getElementById("tankNumber").value)
 
       if (!tankData[tankNumber]) {
         alert("Invalid tank number. Please enter a valid tank number.");
@@ -68,3 +46,7 @@ function toggleMenu() {
 
       document.getElementById("results").style.display = "block";
     }
+
+    // Attach functions to the global window object
+      window.calculateFoam = calculateFoam;
+      window.toggleMenu = toggleMenu;
